Centralized APIs routes
Se movió todo el contenido de los archivos de distintas rutas de las APIs a un solo archivo.
