+++ Routers/Router_API.js
@@ -1,8 +1,16 @@
+require('dotenv').config();
+const API_USER = require('express').Router();
 const API_Carrito = require('express').Router();
+const API_Producto = require('express').Router();
 const controller = require('./Controllers/API.controller');
 const auth = require('./auth/auth');
+const multer = require('multer');
+const { storage } = require('../utils/MulterStorage');
+const upload = multer({ storage });
 
-/* ============ API CARRITO ============= */
+/* =========== ROUTES =========== */
+
+/* API CARRITO */
 
 API_Carrito.use(auth.validateAdmin)// Middleware para validar en todas las rutas que el usuario sea administrador
 
@@ -42,4 +50,39 @@ API_Carrito.delete('/producto/all/:id_prod', auth.isLogged, controller.cart.dele
 
 API_Carrito.delete('/producto/:id_prod', auth.isLogged, controller.cart.decreaseQuantityOnCart);
 
-module.exports = { API_Carrito };
+
+/* API PRODUCTOS */
+
+
+API_Producto.use(auth.validateAdmin)// Middleware para validar en todas las rutas que el usuario sea administrador
+
+API_Producto.get('/', controller.products.allProducts);
+
+API_Producto.get('/:id', controller.products.byProductId);
+
+API_Producto.post('/', controller.products.createProduct);
+
+API_Producto.put('/:id', controller.products.updateProduct);
+
+API_Producto.delete('/:id', controller.products.deleteProduct);
+
+
+/* API USER */
+
+API_USER.get('/', auth.isLogged, controller.user.userInfo);
+
+API_USER.get('/all', controller.user.allUsers);
+
+API_USER.get('/:id', controller.user.Info);
+
+API_USER.post('/buy', auth.isLogged, controller.user.userPurchase);
+
+API_USER.put('/update', auth.isLogged, upload.single('avatar'), controller.user.user_update);
+
+API_USER.delete('/', auth.isLogged, controller.user.deleteUser);
+
+module.exports = {
+    API_Carrito,
+    API_Producto,
+    API_USER
+};
