+++ server.js
@@ -10,9 +10,7 @@ const { Server } = require('http');
 const { socket } = require('./utils/SocketIO');
 const { engine } = require('express-handlebars');
 const { allowInsecurePrototypeAccess } = require('@handlebars/allow-prototype-access');
-const { API_Carrito } = require('./Routers/API_Carrito');
-const { API_Producto } = require('./Routers/API_Producto');
-const { API_USER } = require('./Routers/API_USER');
+const { API_Carrito, API_Producto, API_USER } = require('./Routers/Router_API');
 const { Route: USER_FRONT } = require('./Routers/Router_USER');
 const ms = require('ms');
 const app = express();
